+++ tfx/dsl/component/experimental/annotations_test.py
@@ -25,7 +25,7 @@ from typing import Text
 
 import tensorflow as tf
 
-from tfx.components.base import annotations
+from tfx.dsl.component.experimental import annotations
 from tfx.types import artifact
 from tfx.types import standard_artifacts
