+++ tfx/dsl/component/experimental/decorators_test.py
@@ -31,12 +31,12 @@ import tensorflow as tf
 from tfx import types
 from tfx.components.base import base_executor
 from tfx.components.base import executor_spec
-from tfx.components.base.annotations import InputArtifact
-from tfx.components.base.annotations import OutputArtifact
-from tfx.components.base.annotations import OutputDict
-from tfx.components.base.annotations import Parameter
-from tfx.components.base.decorators import _SimpleComponent
-from tfx.components.base.decorators import component
+from tfx.dsl.component.experimental.annotations import InputArtifact
+from tfx.dsl.component.experimental.annotations import OutputArtifact
+from tfx.dsl.component.experimental.annotations import OutputDict
+from tfx.dsl.component.experimental.annotations import Parameter
+from tfx.dsl.component.experimental.decorators import _SimpleComponent
+from tfx.dsl.component.experimental.decorators import component
 from tfx.orchestration import metadata
 from tfx.orchestration import pipeline
 from tfx.orchestration.beam import beam_dag_runner
